WIP: Started working in adding naivebayes classifier to the app

+++ web/src/test/kotlin/org/metplus/cruncher/web/controller/ResumeControllerTest.kt
@@ -102,7 +102,7 @@ internal class ResumeControllerTest(@Autowired private val mvc: MockMvc) {
                 userId = "asdasdasd",
                 filename = "line_with_bold.pdf",
                 fileType = "pdf",
-                cruncherData = CruncherMetaData(mutableMapOf<String, Long>() as HashMap<String, Long>)
+                cruncherData = CruncherMetaData(mutableMapOf<String, Double>() as HashMap<String, Double>)
         ))
         resumeFileRepository.save(ResumeFile(
                 userId = "asdasdasd",
